Fix some more stuff while breaking some things, oops
Split out CLI stuffs for both DungeonRandomizer.py and Gui.py
Try to fix Generation button
Ended up breaking multiworld stuff?
